Change IEnumerable to ICollection for navigation properties in entities Add IExpensesService interface Add Expenses and ExpensesList models

+++ ExpensesCounter.Web.DAL/Entities/Expense.cs
@@ -2,11 +2,11 @@ namespace ExpensesCounter.Web.DAL.Entities
 {
     public class Expense : BaseEntity
     {
-        public int          ExpenseListId { get; set; }
-        public ExpensesList ExpenseList   { get; set; }
+        public int ExpenseListId { get; set; }
+        public ExpensesList ExpenseList { get; set; }
 
         public string Comment { get; set; }
 
-        public double Count { get; set; }
+        public decimal Count { get; set; }
     }
 }
